+++ keta.js
@@ -108,7 +108,7 @@ angular.module('keta', [
  *         // the directive sets default links that can be overwritten by the keys of this object
  *         $scope.links = {
  *             ALL_APPS: '/#/applications/',
- *             ALL_ENERGY_MANAGERS: '/#/devices?deviceClass=com.kiwigrid.devices.EnergyManager',
+ *             ALL_ENERGY_MANAGERS: '?deviceClass=com.kiwigrid.devices.EnergyManager/#/devices',
  *             APP_ROOT: '/#/landing-page',
  *             USER_PROFILE: '/#/users/profile',
  *             USER_LOGOUT: '/#/users/logout'
@@ -487,29 +487,10 @@ angular.module('keta.directives.AppBar',
 				// get access token
 				var accessToken = AccessToken.decode(AccessToken.get());
 
-				/**
-				 * extend entry uri by current user id
-				 * @param {string} appEntryUri app entry uri to extend
-				 * @returns {string} extended app entry uri
-				 */
-				var extendEntryUriByCurrentUser = function extendEntryUriByCurrentUser(appEntryUri) {
-
-					if (CommonUtils.doesPropertyExist(accessToken, 'user_id')) {
-						if (appEntryUri.indexOf('?') === -1) {
-							appEntryUri += '?';
-						} else {
-							appEntryUri += '&';
-						}
-						appEntryUri += 'userId=' + accessToken.user_id;
-					}
-
-					return appEntryUri;
-				};
-
 				var defaultLinks = {
 					ALL_APPS: null,
 					ALL_ENERGY_MANAGERS: null,
-					APP_ROOT: extendEntryUriByCurrentUser('#/'),
+					APP_ROOT: CommonUtils.addUrlParameter('/', 'userId', accessToken.user_id),
 					USER_PROFILE: null,
 					USER_LOGOUT: null
 				};
@@ -577,24 +558,10 @@ angular.module('keta.directives.AppBar',
 									var link = document.createElement('a');
 									link.href = entryUri;
 
-									var linkProtocol =
-										link.protocol +
-										(link.protocol[link.protocol.length - 1] === ':' ?
-											'//' : '://');
-
-									var linkPort =
-										link.port !== '' && link.port !== '0' ?
-											':' + link.port : '';
-
-									// workaround for internet explorer not having "origin" property
-									var linkOrigin =
-										angular.isDefined(link.origin) ?
-											link.origin :
-											linkProtocol + link.hostname + linkPort;
-
 									if (entryUri !== null) {
 										scope.rootApp = {};
-										scope.rootApp.link = extendEntryUriByCurrentUser(entryUri);
+										scope.rootApp.link =
+											CommonUtils.addUrlParameter(entryUri, 'userId', accessToken.user_id);
 										scope.rootApp.name = name;
 										scope.worlds.unshift({
 											name: 'Desktop',
@@ -604,20 +571,24 @@ angular.module('keta.directives.AppBar',
 
 									scope.links.ALL_APPS = angular.isString(scope.links.ALL_APPS) ?
 										scope.links.ALL_APPS :
-										extendEntryUriByCurrentUser(linkOrigin + '#/applications' + link.search);
+										CommonUtils.addUrlParameter(
+											entryUri + '#/applications', 'userId', accessToken.user_id
+										);
 
 									scope.links.USER_PROFILE = angular.isString(scope.links.USER_PROFILE) ?
 										scope.links.USER_PROFILE :
-										extendEntryUriByCurrentUser(linkOrigin + '#/user' + link.search);
+										CommonUtils.addUrlParameter(
+											entryUri + '#/user', 'userId', accessToken.user_id
+										);
 
 									if (!angular.isString(scope.links.ALL_ENERGY_MANAGERS)) {
-										var linkSearch = link.search;
-										linkSearch += linkSearch === '' ?
-											'?deviceClass=com.kiwigrid.devices.em.EnergyManager' :
-											'&deviceClass=com.kiwigrid.devices.em.EnergyManager';
+										var allManagersUri =
+											CommonUtils.addUrlParameter(
+												entryUri, 'deviceClass', 'com.kiwigrid.devices.em.EnergyManager'
+											);
 										scope.links.ALL_ENERGY_MANAGERS =
-											extendEntryUriByCurrentUser(
-												linkOrigin + '#/devices' + linkSearch
+											CommonUtils.addUrlParameter(
+												allManagersUri + '#/devices', 'userId', accessToken.user_id
 											);
 									}
 
@@ -8213,6 +8184,50 @@ angular.module('keta.utils.Common', [])
 				label = labels[FALLBACK_LOCALE][key];
 			}
 			return label;
+		};
+
+		/**
+		 * @name getLabelByLocale
+		 * @function
+		 * @description
+		 * <p>
+		 *   Add or modify a parameter in given URL. It maintains the correct order or URL parts.
+		 * </p>
+		 * @param {string} uri uri to modify
+		 * @param {string} param parameter to modify
+		 * @param {string} value value to set parameter to
+		 * @returns {string} modified url
+		 */
+		factory.addUrlParameter = function addUrlParameter(uri, param, value) {
+
+			// using a positive lookahead (?=\=) to find the
+			// given parameter, preceded by a ? or &, and followed
+			// by a = with a value after than (using a non-greedy selector)
+			// and then followed by a & or the end of the string
+			var val = new RegExp('(\\?|\\&)' + param + '=.*?(?=(&|$))');
+			var parts = uri.toString().split('#');
+			var url = parts[0];
+			var hash = parts[1] || false;
+			var queryString = /\?.+$/;
+			var newURL = url;
+
+			// check if the parameter exists
+			if (val.test(url)) {
+				// if it does, replace it, using the captured group
+				// to determine & or ? at the beginning
+				newURL = url.replace(val, '$1' + param + '=' + value);
+			} else {
+				// otherwise, if there is a query string at all
+				// add the param to the end of it or
+				// if there's no query string, add one
+				newURL = url + (queryString.test(url) ? '&' : '?') + param + '=' + value;
+			}
+
+			if (hash) {
+				newURL += '#' + hash;
+			}
+
+			return newURL;
 		};
 
 		return factory;